btrfs-progs: offline filesystem resize feature
This patch introduces the ability to resize a btrfs filesystem while it
is not mounted via a new `--offline` flag. Currently only increasing the
size of the filesystem is supported, though I believe it would be
possible
to implement shrinking the filesystem to the end of the last device
extent.

This is a more general, and hopefully more useful, solution to the
problem
I was trying to solve with the
("btrfs-progs: add slack space for mkfs --shrink") patch. This patch
should enable users to resize a filesystem without the higher
capabilities needed for mounting a filesystem.

Signed-off-by: Leo Martins <loemra.dev@gmail.com>
---
Changelog:

v1->v2:
- use chunk root instead of fs root
- fix offline resize error handling
- fix variable declarations to not have newlines
- fix parameter list to have more important arguments first